feat: Make Rating `Copy`
The original idea was to not make it too easy to copy Ratings around
lest one gets confused which ones are the current ones, and 128 bits is
on the border of things that should be trivially copyable, but the
usability win is worth it here, I think.

Thanks to @LFalch for the suggestion.